Change to float field when API call inference

+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_part_detections/api/serializers.py
@@ -160,5 +160,5 @@ class UpdateCamBodySerializer(serializers.Serializer):
         recording_duration = serializers.IntegerField()
 
     lva_mode = serializers.ChoiceField(INFERENCE_PROTOCOL_CHOICES)
-    fps = serializers.IntegerField()
+    fps = serializers.FloatField()
     cameras = CameraItem(many=True)
